% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Extract data from csv file produced by Onset HOBOware}
\usage{
get_data(my_file)
}
\arguments{
\item{my_file}{An \code{import_file} object.}
}
\value{
This function returns a \code{\link[tibble:tibble]{tibble}} of standardized
data. The variables produced by this function include:

\itemize{
\item{\strong{RID}}{The unique record ID. The record ID
\code{\link[base:paste]{paste}}'s DateTime, PlotID, and Element to
create a unique record ID.}
\item{\strong{FileName}}{The name of the file the data came from.}
\item{\strong{PlotID}}{The unique plot identification number (e.g., A03 orI06).}
\item{\strong{DateTime}}{The date and time the data were collected at.}
\item{\strong{Element}}{The element the data represent. TEMP is temperature, RH is
relative humidity, and PRCP is precipitation.}
\item{\strong{Value}}{The data value recorded by the data logger.}
}
}
\description{
This function extracts data from a comma delimited (csv) file produced by
HOBOware. This function uses the list, or object, produced from
\code{\link{import_file}}.
}
\details{
This function extracts the data from the list produced by import_file() then
standardizes the data structure into six columns.
}
\examples{
\dontrun{
library("rainDanceR")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "rainDanceR"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read file into R
my_file <- import_file(file_list[1])

# Extract data from file
get_data(my_file)
}

}
\seealso{
\code{\link{import_file}}, \code{\link{get_details}}
}
