% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details}
\alias{get_details}
\title{Extract metadata from csv file produced by Onset HOBOware}
\usage{
get_details(my_file, my_data)
}
\arguments{
\item{my_file}{An \code{import_file} object.}

\item{my_data}{The data frame produced from \code{get_data}.}
}
\value{
This function returns a \code{\link[tibble:tibble]{tibble}} of metadata on
the. The variables produced by this function include:\preformatted{\\describe\{
    \\item\{\\strong\{FileName\}\}\{ The name of the file the data came from.\}
    \\item\{\\strong\{Details\}\}\{ This variable will varry with the data
        contained in the "Details" colum of the original file. Calculated
        variables listed below.\}
        \\itemize\{
            \\item\{Import Date: A character string of the date the data
                were processed useing lubridate::today().\}
            \\item\{Plot ID: The unique plot identification number (e.g.,
                A03 or I06).\}
            \\item\{Element: The element the data represent. TEMP is
                temperature, RH is relative humidity, and PRCP is
                precipitation.\}
            \\item\{Units: The unit of measurement. Temperature is measured
                in F or C, relative humidity is measured in percent, and
                precipitation is measured in tips of the tipping bucket,
                or events.\}
            \\item\{DateTime (min): The minimum date-time stamp in the data.\}
            \\item\{DateTime (max): The maximum date-time stamp in the data.\}
            \\item\{Records (n): The number of records of data.\}
            \\item\{ConvertFtoC: Logical. Were the temperature data
                converted from F to C? Yes or No.\}
                \}
    \\item\{\\strong\{Value\}\}\{ The value the variable in Details.\}
\}
}
}
\description{
This function extracts metadata from the 'Details' column and raw data
of a comma delimited (csv) file produced by HOBOware. This function uses
the list, or object, produced from \code{\link{import_file}}.
}
\details{
This function produces a data frame of metadata about the sampling event.
Some of the values details are extracted from the "Details" column of the csv
file, some are scraped from the file name, and others are calculated form the
data.
}
\examples{
\dontrun{
library("raindancer")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "raindancer"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read file into R
my_file <- import_file(file_list[1])

# Extract data from file
my_data <- get_data(my_file)

# Extract file details
get_details(my_file, my_data)
}

}
\seealso{
\code{\link{import_file}}, \\code{\link{get_data}}
}
