% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import a csv file produced by Onset HOBOware into R}
\usage{
import_file(
  my_file,
  datestamp_loc = 1,
  plotid_loc = 2,
  plotid_s = 1,
  plotid_e = 3
)
}
\arguments{
\item{my_file}{A character string of the file name. This can include full
directory path.}

\item{datestamp_loc}{A number for the location of the date stamp in the file
name, an index. Default is 1.}

\item{plotid_loc}{A number for the location of the Plot ID in the file name,
an index. Default is 2.}

\item{plotid_s}{A number for the beginning of the Plot ID in the file name,
an index. Default is 1.}

\item{plotid_e}{A number for the end of the Plot ID in the file name, an
index. Default is 3.}
}
\value{
This function returns a list with two components.

\describe{
\item{\strong{file_info}}{This component is a vector that contains the
file name, the date stamp, plot ID, the number of lines to skip to
properly import the data, and the number of columns of data in the raw
file. Varibles listed below:}
\itemize{
\item{filename: The base name of the file being processed.}
\item{datestamp: A date stamp. stripped from the file name.}
\item{plotid: The plot identification number. Stripped from
the file name.}
\item{skip:The number of lines to skip to read in the file
correctly.}
\item{col_n: The number of columns in the raw data file.}}
\item{\strong{raw_file}}{This component is a data frame containing the
raw data and the "Details" column produced by HOBOware.}
}
}
\description{
This function extracts data from the file name and comma delimited (csv)
file produced by HOBOware so the data can be imported into R. This
function produces a list that will be used in subsequent functions
including \code{get_data} and \code{get_details}.
}
\details{
This function extracts variables from the file name and the raw data so the
file can be imported into R. It returns a list with two components. The
first component is a vector containing information about the file. The
second component is a data frame containing the raw data and the
metadata for the sampling event.
}
\examples{
\dontrun{
library("rainDanceR")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "rainDanceR"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read file into R
import_file(file_list[1])
}

}
\seealso{
\code{\link{get_data}}, \code{\link{get_details}}
}
