% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hobo_2008.R
\name{import_hobo_2008}
\alias{import_hobo_2008}
\title{Process a *.csv file produced by Onset HOBOware (2008-2019)}
\usage{
import_hobo_2008(my_file)
}
\arguments{
\item{my_file}{A character string of the file name. This should include the
directory path.}
}
\value{
This function returns a list with three (3) components.

\describe{
\item{\strong{file_info}}{This component is a vector that contains the
file name, the date stamp, plot ID, the number of lines to skip to
properly import the data, the number of columns of data in the
raw file, and the Elements measured..}
\item{\strong{details}}{This component is a data frame of metadata.}
\item{\strong{data_raw}}{This component is a data frame.}
}
}
\description{
This function imports a comma delimited (*.csv) file produced by Onset
HOBOware from Onset data loggers used in NPS Southeast Utah Group parks
from 2008 to 2019. It uses the file name or full path name to produce
a list with four components that contain the file information needed to
import the csv, metadata about the logger and sampling time, and the raw
data.
}
\details{
This function imports the data from a csv file into R and retuns a list
containing the data used to import the file, the metadata, and the
raw data.
}
\examples{
\dontrun{
library("raindancer")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "raindancer"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Import data
import_hobo_2008(file_list[10])
}

}
\seealso{
\code{\link{import_hobo_2020}}
}
