% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_hobo_2008_to_db.R
\name{import_hobo_2008_to_db}
\alias{import_hobo_2008_to_db}
\title{Import Onset Hobo data into Access Database}
\usage{
import_hobo_2008_to_db(
  my_file,
  my_db,
  import_table,
  raw_data_table,
  prcp_data_table,
  temp_rh_data_table,
  details_table,
  verbose = TRUE
)
}
\arguments{
\item{my_file}{A character string of the complete file path of your *.csv
file.}

\item{my_db}{A connected database from \code{\link{RODBC}}.}

\item{import_table}{A character string of the name of the import log table.}

\item{raw_data_table}{A character string of the name of the raw data table.}

\item{prcp_data_table}{A character string of the name of the processed
precipitation data table.}

\item{temp_rh_data_table}{A character string of the name of the processed
temperature and relative humidity data table.}

\item{details_table}{A character string of the name of the logger details
table.}

\item{verbose}{Logical. Default is TRUE. If FALSE, messages are suppressed.}
}
\value{
Data is written to database tables. Objects are not returned.
}
\description{
This function processes data from Onset HOBOware and exports them to a
Microsoft Access Database.
}
\examples{
\dontrun{
library("dataprocessR")

# Connect to DB
my_db <- RODBC::odbcConnectAccess2007("C:/path/to/database.accdb")

# List files
my_dir <- "C:/path/to/data"
file_list <- list.files(my_dir, pattern = ".csv", full.names = TRUE,
                        recursive = FALSE)
# Select file
my_file <- file_list[10]

# Process file and save to database
import_hobo_2008_to_db(my_file = my_file, my_db = my_db,
                       import_table = "tbl_import_log",
                       raw_data_table = "tbl_raw_data",
                       prcp_data_table = "tbl_prcp_data",
                       temp_rh_data_table = "tbl_temp_rh_data",
                       details_table = "tbl_logger_details")
}
}
