% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_wxdat.R
\name{import_wxdat}
\alias{import_wxdat}
\title{Process a *.csv file produced by Onset HOBOware}
\usage{
import_wxdat(this_file, ...)
}
\arguments{
\item{this_file}{A character string of the file name. This can include full
directory path.}

\item{...}{Other arguments to pass to \code{\link{import_file}}. See
\code{\link{import_file}} for more information.}
}
\value{
This function returns a list with four components.

\describe{
\item{\strong{file_info}}{This component is a vector that contains the
file name, the date stamp, plot ID, the number of lines to skip to
properly import the data, the number of columns of data in the
raw file, and the Elements measured. This comonent is the product of
\code{\link{import_file}}.}
\item{\strong{details}}{This component is a data frame of metadata from
\code{\link{get_details}}.}
\item{\strong{data}}{This component is a data frame from
\code{\link{get_data}}.}
\item{\strong{raw_file}}{The raw file that was loaded into R using
\code{\link{import_file}}.}
}
}
\description{
This function combines \code{\link{import_file}}, \code{\link{get_data}}, and
\code{\link{get_details}} to process a comma delimited (*.csv) file
produced by HOBOware. It uses the file name or full path name to produce
a list with three components that contain the file information needed to
import the csv, the raw data, and associated metadata.
}
\details{
This function imports the data from a csv file into R and retuns a list
containing the data used to import the file, the metadata, and the
raw data. This function uses \code{\link{import_file}},
\code{\link{get_data}}, and \code{\link{get_details}} process the file.
}
\examples{
\dontrun{
library("rainDanceR")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "rainDanceR"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Import data
import_wxdat(file_list[1])
}

}
\seealso{
\code{\link{import_file}}, \code{\link{get_data}},
\code{\link{get_details}}
}
