% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sundance.R
\name{sundance}
\alias{sundance}
\title{Summarize temperature data}
\usage{
sundance(my_wxdat)
}
\arguments{
\item{my_wxdat}{An \code{import_wxdat} object.}
}
\value{
This function returns a nine (9) column \code{\link[tibble:tibble]{tibble}}.

\describe{
\item RID: The unique record ID. The record ID
\code{\link[base:paste]{paste}}'s Date and PlotID to  create a unique
record ID.
\item{\strong{PlotID}}{The unique ID number for the long-term monitoring
plot.}
\item{\strong{Date}}{The date the data were recorded.}
\item{\strong{TEMP_mean}}{The mean daily temperature.}
\item{\strong{TMIN}}{The minimum daily temperature.}
\item{\strong{TMAX}}{The maximum daily temperature.}
\item{\strong{n}}{The number of records for that day.}
\item{\strong{TMIN_time}}{The time of daily minimum temperature.}
\item{\strong{TEMP_mean}}{The time of daily maximum temperature.}
}
}
\description{
This function summarizes temperature data from Onset Hobo event loggers. It
uses a list produced from \code{\link{import_wxdat}} and returns a data
frame of temperature data.
}
\details{
This function summarizes temperature data from Onset Hobo event  loggers.\preformatted{This function srtips the first 5 minutes and last 10 minutes of data to
account for field proceedures when downloading the event logger. It is
common practice to trigger an event before downloading and after
launching the logger. To remove these false events there is a routine to
stip the first 5 minutes of the data if the "Launch Time" and "First
Sample Time" are the same. There is no way to determine if the tipping
bucket is triggered before downloading the data, so the last 10 minutes
of every file is stripped before processing.
}
}
\examples{
\dontrun{
library("rainDanceR")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "rainDanceR"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read file into R
my_temp <- import_wxdat(file_list[2])

# Process precipitation data
sundance(my_temp)
}
}
\seealso{
\code{\link{import_wxdat}} to import a csv produced by Onset
HOBOware into R.
}
