% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_wxdat.R
\name{import_wxdat}
\alias{import_wxdat}
\title{Process a *.csv file produced by Onset HOBOware}
\usage{
import_wxdat(this_file, ...)
}
\arguments{
\item{this_file}{A character string of the file name. This can include full
directory path.}

\item{...}{Other arguments to pass to \code{\link{import_file}}. See
\code{\link{import_file}} for more information.}
}
\value{
This function returns a list with three components.

\describe{
\item{\strong{file_info}}{This component is a vector that contains the
file name, the date stamp, plot ID, the number of lines to skip to
properly import the data, and the number of columns of data in the
raw file. This comonent is a component of \code{\link{import_file}}.}
\item{\strong{details}}{This component is a data frame of metadata. See
\code{\link{get_details}} for more details.}
\item{\strong{data}}{This component is a data frame containing the
standardized form of the raw data. See \code{\link{get_data}} for
more details.}
\item{\strong{raw_file}}{The raw, unprocessed or standardized file that
was loaded into R using \code{\link{import_file}}. See
\code{\link{import_file}} for more details.}
}
}
\description{
This function combines \code{\link{import_file}}, \code{\link{get_data}}, and
\code{\link{get_details}} to process a comma delimited (csv) file
produced by HOBOware. It uses the file name or full path name to produce
a list with three components that contain the file information needed to
import the csv, the raw data, and associated metadata.
}
\details{
This function imports the data from a csv file into R and retuns a list
containing the data used to import the file, the metadata, and the
raw data. This function uses \code{\link{import_file}},
\code{\link{get_data}}, and \code{\link{get_details}} process the file.
}
\examples{
\dontrun{
library("dataProcessor")
file.list <- list.files(path = "./inst/raw_data", pattern = ".csv",
                        full.names = TRUE, recursive = FALSE)
import_wxdat(file.list[1])
}

}
\seealso{
\code{\link{import_file}} to import a csv produced by HOBOware,
\code{\link{get_data}} to extract the data, and \code{\link{get_details}}
to extract the metadata.
}
