% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raindance.R
\name{raindance}
\alias{raindance}
\title{Summarize precipitation data}
\usage{
raindance(my_data)
}
\arguments{
\item{my_data}{A data frame with four columns. Typically from
\code{\link{get_data}}. At a minimum, columns must include the following:
\describe{
\item{\strong{Element}}{The element the data represent. TEMP is
temperature, RH is relative humidity, and PRCP is precipitation.}
\item{\strong{PlotID}}{The unique plot identification number (e.g.,
A03 or I06).}
\item{\strong{DateTime}}{The date-time of the measurement.}
\item{\strong{Value}}{The data value of the measurement recorded by
the data logger.}
}}
}
\value{
This function returns a five column \code{\link[tibble:tibble]{tibble}}.

\describe{
\item{\strong{RID}}{The unique record ID. The record ID
\code{\link[base:paste]{paste}}'s DateTime and PlotID to create a
unique record ID.}
\item{\strong{DateTime}}{The date and hour of the data. Hours are the from
0 min 00 sec to 59 min 59 sec. For example 2009-04-15 00:00:00 is
from 00:00:00 to 00:59:59 on April 15, 2009.}
\item{\strong{PlotID}}{The unique ID number for the long-term monitoring
plot.}
\item{\strong{PRCP_mm}}{The recorded total precipitation for that hour in
millimeters.}
\item{\strong{Tips}}{The number of tips recorded for that hour.}
\item{\strong{MaxTips_min}}{The maximum tips per minute recorded. This is
inteded to calculate intensity of precipitation event.}
}
}
\description{
This function summarizes precipitation data from Onset event loggers
used in Onset tipping bucket precipitation gauges.
}
\details{
This function summarizes precipitation data from Onset event loggers
used in Onset tipping bucket precipitation gauges. It uses a data frame
typically produced from \code{\link{get_data}} and returns a data frame of
hourly precipitation totals, number of tips per hour, and maximum tips
per minute.
}
\examples{
\dontrun{
library("rainDanceR")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "rainDanceR"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read file into R
my_prcp <- import_wxdat(file_list[1])$data_raw

# Process precipitation data
raindance(my_prcp)
}

}
\seealso{
\code{\link{get_data}}, \code{\link{import_wxdat}}
}
