% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raindance.R
\name{raindance}
\alias{raindance}
\title{Summarize precipitation data}
\usage{
raindance(my_data)
}
\arguments{
\item{my_data}{A data frame with from \code{\link{import_hobo_2008}} or
\code{\link{import_hobo_2020}}. If not from these functions, the columns
must include the following:
\describe{
\item{\strong{Element}}{ The element the data represent. Onset
tipping bucket evnet loggers record an event when 0.254 mm of
precipitation tips the bucket.}
\item{\strong{PlotID}}{ The unique plot identification number (e.g.,
A03 or I06).}
\item{\strong{DateTime}}{ The date-time of the precipitation event}
\item{\strong{Value}}{ Event data value recorded by the data logger.}
}}
}
\value{
This function returns a \code{\link[tibble:tibble]{tibble}}.

\describe{
\item{\strong{DateTime}}{ The date and hour of the data. Hours are the from
0 min 00 sec to 59 min 59 sec. For example 2009-04-15 00:00:00 is
from 00:00:00 to 00:59:59 on April 15, 2009.}
\item{\strong{PlotID}}{ The unique ID number for the long-term monitoring
plot.}
\item{\strong{PRCP_mm}}{ The recorded total precipitation for that hour in
millimeters.}
\item{\strong{Tips}}{ The number of tips recorded for that hour.}
\item{\strong{MaxTips_min}}{ The maximum tips per minute recorded. This is
intended to calculate intensity of precipitation event.}
}
}
\description{
This function summarizes precipitation data from Onset event loggers
used in Onset tipping bucket precipitation gauges.
}
\details{
This function summarizes precipitation data from Onset event loggers
used in Onset tipping bucket precipitation gauges. It uses the data_raw
data frame produced from \code{\link{import_hobo_2008}} or
\code{\link{import_hobo_2020}} and returns a data frame of hourly
precipitation totals, number of tips per hour, and maximum tips per
minute.
}
\examples{
\dontrun{
library("raindancer")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "raindancer"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read data into R
my_prcp <- import_hobo_2008(file_list[5])$data_raw

# Process precipitation data
raindance(my_prcp)
}

}
\seealso{
\code{\link{import_hobo_2008}}, \code{\link{import_hobo_2020}}
}
