% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raindance.R
\name{raindance}
\alias{raindance}
\title{Summarize precipitation data}
\usage{
raindance(my_wxdat)
}
\arguments{
\item{my_wxdat}{An \code{import_wxdat} object.}
}
\value{
This function returns a five column \code{\link[tibble:tibble]{tibble}}.

\describe{
\item{\strong{RID}}{The unique record ID. The record ID
\code{\link[base:paste]{paste}}'s DateTime and PlotID to create a
unique record ID.}
\item{\strong{DateTime}}{The date and hour of the data. Hours are the from
0 min 00 sec to 59 min 59 sec. For example 2009-04-15 00:00:00 is
from 00:00:00 to 00:59:59 on April 15, 2009.}
\item{\strong{PlotID}}{The unique ID number for the long-term monitoring
plot.}
\item{\strong{PRCP_mm}}{The recorded total precipitation for that hour in
millimeters.}
\item{\strong{Tips}}{The number of tips recorded for that hour.}
\item{\strong{MaxTips_min}}{The maximum tips per minute recorded. This is
inteded to calculate intensity of precipitation event.}
}
}
\description{
This function summarizes precipitation data from Onset Hobo event loggers
used in Onset tipping bucket precipitation gauges. It uses a list
produced from \code{\link{import_wxdat}} and returns a data frame of
hourly precipitation totals, number of tips per hour, and maximum tips
per minute.
}
\details{
This function summarizes precipitation data from Onset Hobo event  loggers
used in Onset tipping bucket precipitation gauges.\preformatted{This function srtips the first 5 minutes and last 10 minutes of data to
account for field proceedures when downloading the event logger. It is
common practice to trigger an event before downloading and after
launch_timeing the logger. To remove these false events there is a routine to
stip the first 5 minutes of the data if the "launch_time Time" and "First
Sample Time" are the same. There is no way to determine if the tipping
bucket is triggered before downloading the data, so the last 10 minutes
of every file is stripped before processing.
}
}
\examples{
\dontrun{
library("rainDanceR")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "rainDanceR"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read file into R
my_prcp <- import_wxdat(file_list[1])

# Process precipitation data
raindance(my_prcp)
}

}
\seealso{
\code{\link{import_wxdat}}
}
