% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhdance.R
\name{rhdance}
\alias{rhdance}
\title{Summarize relative humidity data}
\usage{
rhdance(my_data)
}
\arguments{
\item{my_data}{A data frame with four columns. Typically from
\code{\link{get_data}}. At a minimum, columns must include the following:
\describe{
\item{\strong{Element}}{The element the data represent. TEMP is
temperature, RH is relative humidity, and PRCP is precipitation.}
\item{\strong{PlotID}}{The unique plot identification number (e.g.,
A03 or I06).}
\item{\strong{DateTime}}{The date-time of the measurement.}
\item{\strong{Value}}{The data value of the measurement recorded by
the data logger.}
}}
}
\value{
This function returns a nine (9) column \code{\link[tibble:tibble]{tibble}}.

\describe{
\item{\strong{RID}}{The unique record ID. The record ID
\code{\link[base:paste]{paste}}'s Date and PlotID to  create a unique
record ID.}
\item{\strong{PlotID}}{The unique ID number for the long-term monitoring
plot.}
\item{\strong{Date}}{The date the data were recorded.}
\item{\strong{RH_mean}}{The mean daily RH}
\item{\strong{RHMIN}}{The minimum daily RH}
\item{\strong{RHMAX}}{The maximum daily RH}
\item{\strong{n}}{The number of records for that day.}
\item{\strong{RHMIN_time}}{The time of minimum daily RH}
\item{\strong{RHMAX_time}}{The time of maximum daily RH}
}
}
\description{
This function summarizes relative humidity (RH) data from Onset loggers.
}
\details{
This function summarizes RH data from Onset Hobo loggers. It uses a list
produced from \code{\link{import_wxdat}} and returns a data frame of
summarized RH data.
}
\examples{
\dontrun{
library("rainDanceR")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "rainDanceR"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read file into R
my_rh <- import_wxdat(file_list[12])$data

# Process precipitation data
rhdance(my_rh)
}
}
\seealso{
\code{\link{get_data}}, \code{\link{import_wxdat}}
}
