% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sundance.R
\name{sundance}
\alias{sundance}
\title{Summarize temperature data}
\usage{
sundance(my_data)
}
\arguments{
\item{my_data}{A data frame with from \code{\link{import_hobo_2008}} or
\code{\link{import_hobo_2020}}. If not from these functions, the columns
must include the following:
\describe{
\item{\strong{Element}}{ The element the data represent. TEMP is
temperature or RH is relative humidity.}
\item{\strong{PlotID}}{ The unique plot identification number (e.g.,
A03 or I06).}
\item{\strong{DateTime}}{ The date-time of the measurement.}
\item{\strong{Value}}{ The data value of the measurement recorded by
the data logger.}
\item{\strong{Units}}{ The unit of the measurement.}
}}
}
\value{
This function returns a \code{\link[tibble:tibble]{tibble}}.

\describe{
\item{\strong{PlotID}}{ The unique ID number for the long-term monitoring
plot.}
\item{\strong{Date}}{ The date the data were recorded.}
\item{\strong{TEMP_mean}}{ The mean daily temperature.}
\item{\strong{TMIN}}{ The minimum daily temperature.}
\item{\strong{TMAX}}{ The maximum daily temperature.}
\item{\strong{n}}{ The number of records for that day.}
\item{\strong{TMIN_time}}{ The time of minimum daily temperature.}
\item{\strong{TEMP_mean}}{ The time of maximum daily temperature.}
\item{\strong{Units}}{ The unit of the measurement.}
}
}
\description{
This function summarizes temperature and relative humidity data from Onset
data loggers.
}
\details{
This function summarizes temperature and relative humidity data from Onset
loggers. It uses the data_raw data frame produced from
\code{\link{import_hobo_2008}} ro \code{\link{import_hobo_2020}} and
returns a data frame of summarized temperature or relative humidity data.
}
\examples{
\dontrun{
library("raindancer")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "raindancer"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Temperature data
my_temp <- import_hobo_2008(file_list[4])$data_raw
sundance(my_temp)

# Relative humidity data
my_rh <- import_hobo_2020(file_list[10])$data_raw
sundance(my_rh)
}
}
\seealso{
\code{\link{import_hobo_2008}}, \code{\link{import_hobo_2020}}
}
