% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sundance.R
\name{sundance}
\alias{sundance}
\title{Summarize temperature data}
\usage{
sundance(my_data)
}
\arguments{
\item{my_data}{A data frame with four columns. Typically from
\code{\link{get_data}}. Columns must include the following:
\describe{
\item{\strong{Element}}{ The element the data represent. TEMP is
temperature, RH is relative humidity, and PRCP is precipitation.}
\item{\strong{PlotID}}{ The unique plot identification number (e.g.,
A03 or I06).}
\item{\strong{DateTime}}{ The date-time of the measurement.}
\item{\strong{Value}}{ The data value of the measurement recorded by
the data logger.}
\item{\strong{Units}}{ The unit of the measurement.}
}}
}
\value{
This function returns a nine (9) column \code{\link[tibble:tibble]{tibble}}.

\describe{
\item{\strong{PlotID}}{ The unique ID number for the long-term monitoring
plot.}
\item{\strong{Date}}{ The date the data were recorded.}
\item{\strong{TEMP_mean}}{ The mean daily temperature.}
\item{\strong{TMIN}}{ The minimum daily temperature.}
\item{\strong{TMAX}}{ The maximum daily temperature.}
\item{\strong{n}}{ The number of records for that day.}
\item{\strong{TMIN_time}}{ The time of minimum daily temperature.}
\item{\strong{TEMP_mean}}{ The time of maximum daily temperature.}
\item{\strong{Units}}{ The unit of the measurement.}
}
}
\description{
This function summarizes temperature data from Onset loggers.
}
\details{
This function summarizes temperature data from Onset loggers. It
uses a list produced from \code{\link{import_wxdat}} and returns a data
frame of summarized temperature data.
}
\examples{
\dontrun{
library("raindancer")

# Generate list of files
file_list <- list.files(path = system.file("extdata", package = "raindancer"),
                        pattern = ".csv", full.names = TRUE, recursive = FALSE)

# Read file into R
my_temp <- import_wxdat(file_list[10])$data_raw

# Process precipitation data
sundance(my_temp)
}
}
\seealso{
\code{\link{get_data}}, \code{\link{import_wxdat}}
}
